split reducers responsibilities
