+++ src/index.js
@@ -4,12 +4,10 @@ import {createStore, applyMiddleware} from 'redux';
 import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 import {composeWithDevTools} from 'redux-devtools-extension';
-import {reducer} from './store/reducers/reducer';
+import reducer from './store/reducers/reducer';
 import App from './components/app/app';
 import {createAPI} from './api';
 import DataOperation from './store/operations/data-operation';
-// import questions from './mocks/questions';
-// import settings from './mocks/settings';
 
 const api = createAPI(() => {});
 const store = createStore(
