+++ src/store/reducers/dataReducer/dataReducer.js
@@ -1,0 +1,7 @@
+import questionsReducer from './questionsReducer/questionsReducer';
+
+const dataReducer = (dataState, action) => ({
+  questions: questionsReducer(dataState.questions, action)
+});
+
+export default dataReducer;
