+++ src/store/reducers/gameReducer/stepReducer/stepReducer.test.js
