+++ src/store/reducers/name-space.js
@@ -1,0 +1,8 @@
+const NameSpace = {
+  DATA: `DATA`,
+  GAME: `GAME`,
+  USER: `USER`
+};
+
+export default NameSpace;
+
