+++ src/store/reducers/userReducer/userReducer.js
@@ -1,0 +1,3 @@
+const userReduder = (userState) => (userState);
+
+export default userReduder;
